data-ingestion/requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3

# Data APIs
requests==2.31.0

# Utilities
pyyaml==6.0.1
click==8.1.7
loguru==0.7.0

# Google Cloud (compatible versions)
google-cloud-storage==2.10.0
google-auth==2.22.0
